Move AsciiString helpers to utilities so I can reuse them all over
