+++ src/EnergyPlus.hpp
@@ -15,15 +15,5 @@ void runEnergyPlus(int argc, const char* argv[], ftxui::Sender<std::string>* sen
 
 bool validateFileType(const std::filesystem::path& filePath);
 
-inline std::string ascii_to_lower_copy(std::string_view input) {
-  std::string result{input};
-  constexpr auto to_lower_diff = 'a' - 'A';
-  for (auto& c : result) {
-    if (c >= 'A' && c <= 'Z') {
-      c += to_lower_diff;
-    }
-  }
-  return result;
-}
 }  // namespace epcli
 #endif  // ENERGYPLUS_HPP
